requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
Pillow==10.1.0
telethon==1.33.1
asyncio==3.4.3